Add bindings to the Elligator Swift module
The Elligator Swift encoding allows representing a valid group elemnt as
a indistinguishable from random field element. This reduces the chance
for a passive attacker to find out about key exchanges where a point is
  sent unencrypted. This is part of the BIP324 effort.

This commit adds ffi-bindings arround libsecp's API for Elligator Swift,
and introduces a safe abstraction over it.